+++ packages/manager/apps/container/src/core/sso.ts
@@ -1,8 +1,12 @@
 export const initSso = (): void => {
   window.addEventListener(
     'message',
-    ({ data }) => {
-      if (data && data.id === 'ovh-auth-redirect') {
+    (event) => {
+      const { data } = event;
+      if (
+        data?.id === 'ovh-auth-redirect' &&
+        event.origin === window.location.origin
+      ) {
         window.location.assign(data.url);
       }
     },