+++ packages/manager/modules/pci/src/projects/project/storages/components/user-credential-banner/component.js
@@ -4,9 +4,11 @@ export default {
   bindings: {
     user: '<',
     credential: '<',
+    fields: '<',
     fieldSecretKeyLabel: '@',
     containerName: '<',
     role: '<',
+    onCopyFieldClick: '&',
   },
   template,
 };