+++ packages/manager/apps/public-cloud/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.24.21](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.20...@ovh-ux/manager-public-cloud@4.24.21) (2023-05-22)
+
+**Note:** Version bump only for package @ovh-ux/manager-public-cloud
+
+
+
+
+
 ## [4.24.20](https://github.com/ovh/manager/compare/@ovh-ux/manager-public-cloud@4.24.19...@ovh-ux/manager-public-cloud@4.24.20) (2023-05-15)
 
 **Note:** Version bump only for package @ovh-ux/manager-public-cloud